+++ crates/mimic/src/data/types.rs
@@ -1,42 +1,6 @@
+use crate::types::prim::Key;
 use candid::CandidType;
-use derive_more::{Deref, DerefMut};
-use icu::impl_storable_unbounded;
 use serde::{Deserialize, Serialize};
-
-///
-/// CompositeKey
-///
-
-#[derive(
-    CandidType,
-    Clone,
-    Debug,
-    Default,
-    Deref,
-    DerefMut,
-    Eq,
-    PartialEq,
-    Hash,
-    Ord,
-    PartialOrd,
-    Serialize,
-    Deserialize,
-)]
-pub struct CompositeKey(pub Vec<String>);
-
-impl<S: ToString> From<Vec<S>> for CompositeKey {
-    fn from(v: Vec<S>) -> Self {
-        Self(v.into_iter().map(|x| x.to_string()).collect())
-    }
-}
-
-impl<S: ToString> From<&[S]> for CompositeKey {
-    fn from(v: &[S]) -> Self {
-        Self(v.iter().map(|x| x.to_string()).collect())
-    }
-}
-
-impl_storable_unbounded!(CompositeKey);
 
 ///
 /// Selector
@@ -55,10 +19,10 @@ pub enum Selector {
     #[default]
     All,
     Only,
-    One(CompositeKey),
-    Many(Vec<CompositeKey>),
-    Prefix(CompositeKey),
-    Range(CompositeKey, CompositeKey),
+    One(Key),
+    Many(Vec<Key>),
+    Prefix(Key),
+    Range(Key, Key),
 }
 
 ///