+++ crates/mimic/src/types/prim/mod.rs
@@ -1,18 +1,18 @@
 mod blob;
 mod decimal;
 mod int;
+mod key;
 mod nat;
 mod principal;
-mod relation;
 mod ulid;
 mod unit;
 
 pub use blob::Blob;
 pub use decimal::Decimal;
 pub use int::Int;
+pub use key::{Key, KeySet};
 pub use nat::Nat;
 pub use principal::Principal;
-pub use relation::{Relation, RelationSet};
 pub use ulid::Ulid;
 pub use unit::Unit;